+++ build_variables.bzl
@@ -105,23 +105,6 @@ jit_core_sources = [
 core_sources_common = [
     "torch/csrc/autograd/autograd_meta.cpp",
     "torch/csrc/autograd/forward_grad.cpp",
-    "torch/csrc/autograd/anomaly_mode.cpp",
-    "torch/csrc/autograd/autograd.cpp",
-    "torch/csrc/autograd/autograd_not_implemented_fallback.cpp",
-    "torch/csrc/autograd/cpp_hook.cpp",
-    "torch/csrc/autograd/custom_function.cpp",
-    "torch/csrc/autograd/engine.cpp",
-    "torch/csrc/autograd/function.cpp",
-    "torch/csrc/autograd/functions/accumulate_grad.cpp",
-    "torch/csrc/autograd/functions/basic_ops.cpp",
-    "torch/csrc/autograd/functions/tensor.cpp",
-    "torch/csrc/autograd/functions/utils.cpp",
-    "torch/csrc/autograd/input_buffer.cpp",
-    "torch/csrc/autograd/record_function_ops.cpp",
-    "torch/csrc/autograd/saved_variable.cpp",
-    "torch/csrc/autograd/variable.cpp",
-    "torch/csrc/autograd/VariableFallbackKernel.cpp",
-    "torch/csrc/autograd/utils/warnings.cpp",
     "torch/csrc/jit/frontend/edit_distance.cpp",
     "torch/csrc/jit/mobile/compatibility/runtime_compatibility.cpp",
     "torch/csrc/jit/mobile/type_parser.cpp",
@@ -171,6 +154,22 @@ libtorch_edge_profiler_sources = libtorch_profiler_sources + [
 ]
 
 core_trainer_sources = [
+    "torch/csrc/autograd/anomaly_mode.cpp",
+    "torch/csrc/autograd/autograd.cpp",
+    "torch/csrc/autograd/autograd_not_implemented_fallback.cpp",
+    "torch/csrc/autograd/cpp_hook.cpp",
+    "torch/csrc/autograd/custom_function.cpp",
+    "torch/csrc/autograd/engine.cpp",
+    "torch/csrc/autograd/function.cpp",
+    "torch/csrc/autograd/functions/accumulate_grad.cpp",
+    "torch/csrc/autograd/functions/basic_ops.cpp",
+    "torch/csrc/autograd/functions/tensor.cpp",
+    "torch/csrc/autograd/functions/utils.cpp",
+    "torch/csrc/autograd/input_buffer.cpp",
+    "torch/csrc/autograd/record_function_ops.cpp",
+    "torch/csrc/autograd/saved_variable.cpp",
+    "torch/csrc/autograd/variable.cpp",
+    "torch/csrc/autograd/utils/warnings.cpp",
     "torch/csrc/autograd/jit_decomp_interface.cpp",
     "torch/csrc/jit/frontend/name_mangler.cpp",
     "torch/csrc/jit/ir/type_hashing.cpp",
@@ -1021,6 +1020,7 @@ aten_cpu_source_non_codegen_list = [
     "aten/src/ATen/core/List.cpp",
     "aten/src/ATen/core/NamedTensor.cpp",
     "aten/src/ATen/core/Tensor.cpp",
+    "aten/src/ATen/core/VariableFallbackKernel.cpp",
     "aten/src/ATen/core/VariableHooksInterface.cpp",
     "aten/src/ATen/core/Vitals.cpp",
     "aten/src/ATen/core/boxing/KernelFunction.cpp",
